refactor: add separate optional props instead of union type, split cluttered func into organized dir & removed dual registration (registers only under import name, not resolved path)
