+++ lib/utils/parse-tsconfig-paths/index.ts
@@ -1,0 +1,4 @@
+export { parseTsconfigPaths } from "./parse-tsconfig-paths"
+export { resolveTsconfigPath } from "./resolve-tsconfig-path"
+export { removeJsonComments } from "./remove-json-comments"
+export type { TsconfigPaths } from "./types"
