+++ lib/utils/parse-tsconfig-paths/types.ts
@@ -1,0 +1,4 @@
+export interface TsconfigPaths {
+  baseUrl?: string
+  paths?: Record<string, string[]>
+}
